moved name to profile screen
